backend/requirements.txt
# ShadowLab backend – Chaos Engineering for AI APIs
# Core API
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# HTTP client for calling target APIs and Gradient AI
httpx>=0.26.0

# Validation and settings
pydantic>=2.5.0
python-dotenv>=1.0.0
